chore(language): Moveclient to LanguageServerController class
